+++ py/DB/Sample.py
@@ -19,7 +19,7 @@ SAMPLE_FREE_COLUMNS = 31
 class Sample(Model):
     # Historical (plural) name of the table
     __tablename__ = 'samples'
-    sampleid = Column(BIGINT, Sequence('seq_samples'), primary_key=True)
+    sampleid = Column(INTEGER, Sequence('seq_samples'), primary_key=True)
     projid = Column(INTEGER, ForeignKey('projects.projid'))
     # i.e. sample_id from TSV
     orig_id = Column(VARCHAR(255))
